Eliminate 'Consensus.Evaluator' module
Merge the `Oscoin.Consensus.Evaluator` module into
`Oscoin.Crypto.Blockchain.Eval`. Also move
`Oscoin.Consensus.Evaluator.Radicle` to `Oscoin.RadicleTx`
