+++ test/Oscoin/Test/Crypto/Blockchain.hs
@@ -4,7 +4,6 @@ module Oscoin.Test.Crypto.Blockchain
 
 import           Oscoin.Prelude
 
-import           Oscoin.Consensus.Evaluator
 import           Oscoin.Crypto.Blockchain.Block
 import           Oscoin.Crypto.Blockchain.Eval
 import           Oscoin.Crypto.Hash
